[refactoring] - Splitting the code in multiple files for code organization
