+++ pokemonApp/PokemonStatType.swift
@@ -1,0 +1,14 @@
+//
+//  PokemonStat.swift
+//  pokemonApp
+//
+//  Created by Anonim on 25.09.2024.
+//
+
+import Foundation
+
+struct PokemonStatType : Identifiable, Hashable{
+    let id = UUID()
+    let name: String
+    let baseStat: Int    
+}
