+++ pokemonApp/pokemonApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct pokemonAppApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            ListView()
         }
     }   
 }
